--- a/jupyter_server/base/handlers.py
+++ b/jupyter_server/base/handlers.py
@@ -31,7 +31,7 @@ from ipython_genutils.py3compat import string_types
 import jupyter_server
 from jupyter_server._tz import utcnow
 from jupyter_server.i18n import combine_translations
-from jupyter_server.utils import is_hidden, url_path_join, url_is_absolute, url_escape
+from jupyter_server.utils import ensure_async, is_hidden, url_path_join, url_is_absolute, url_escape
 from jupyter_server.services.security import csp_report_uri
 
 #-----------------------------------------------------------------------------
@@ -800,13 +800,13 @@ class FilesRedirectHandler(JupyterHandler):
     """Handler for redirecting relative URLs to the /files/ handler"""
 
     @staticmethod
-    def redirect_to_files(self, path):
+    async def redirect_to_files(self, path):
         """make redirect logic a reusable static method
 
         so it can be called from other handlers.
         """
         cm = self.contents_manager
-        if cm.dir_exists(path):
+        if await ensure_async(cm.dir_exists(path)):
             # it's a *directory*, redirect to /tree
             url = url_path_join(self.base_url, 'tree', url_escape(path))
         else:
@@ -814,14 +814,14 @@ class FilesRedirectHandler(JupyterHandler):
             # otherwise, redirect to /files
             parts = path.split('/')
 
-            if not cm.file_exists(path=path) and 'files' in parts:
+            if not await ensure_async(cm.file_exists(path=path)) and 'files' in parts:
                 # redirect without files/ iff it would 404
                 # this preserves pre-2.0-style 'files/' links
                 self.log.warning("Deprecated files/ URL: %s", orig_path)
                 parts.remove('files')
                 path = '/'.join(parts)
 
-            if not cm.file_exists(path=path):
+            if not await ensure_async(cm.file_exists(path=path)):
                 raise web.HTTPError(404)
 
             url = url_path_join(self.base_url, 'files', url_escape(path))