--- a/jupyter_server/services/contents/filemanager.py
+++ b/jupyter_server/services/contents/filemanager.py
@@ -10,7 +10,6 @@ import os
 import shutil
 import stat
 import sys
-import warnings
 import mimetypes
 import nbformat