--- a/setup.py
+++ b/setup.py
@@ -49,6 +49,7 @@ setup_args = dict(
         'terminado>=0.8.3',
         'prometheus_client',
         "pywin32>=1.0 ; sys_platform == 'win32'",
+        "anyio>=2.0.2",
     ],
     extras_require = {
         'test': ['coverage', 'requests',