**Description**
Hi, here's the draft for adding an async option for the contents api. 

So far, I added an async option for each class/mixin below:
- `ContentsManager` and `AsyncContentsManager`
- `Checkpoints` and `AsyncCheckpoints`
- `GenericCheckpointsMixin` and `AsyncGenericCheckpointsMixin`

@kevin-bates and/or @Zsailer I'd appreciate your review and feedback!

**Open Questions**
I added a test for the configuration, but how do we want to test the new classes since most methods are not implemented and left up to the user?

**Checklist**
- [x] Update Contents Api Doc
- [x] Add test for contents configuration
- [x] Implement `AsyncFileContentsManager`
- [x] Update test_ap/test_manager to test both `FileContentsManager` and `AsyncFileContentsManager`
- [x] Rebase with master branch to include public fixture update. 
- [x] Remove python 3.5 support code. 
- [x] Fix code scanning lint errors
- [x] Test new classes with my team's custom `ContentsManager`
- [x] Update docs to include best use cases for the async classes. 
- [x] Fix tests