@kevin-bates 2020-11-03
Thanks @mwakaba2 - this is a good start.

> how do we want to test the new classes since most methods are not implemented and left up to the user?

You're right, most tests leverage the `FileContentsManager` and things should get "concrete" when adding its async form.  At that time, I suspect we could use parameters like in the kernels tests, which run the tests using [both classes](https://github.com/jupyter/jupyter_server/blob/master/tests/services/kernels/test_api.py#L18-L22)...
```
@pytest.fixture(params=["MappingKernelManager", "AsyncMappingKernelManager"])
def argv(request):
    if request.param == "AsyncMappingKernelManager" and sys.version_info < (3, 6):
        pytest.skip("Kernel manager is AsyncMappingKernelManager, Python version < 3.6")
    return ["--ServerApp.kernel_manager_class=jupyter_server.services.kernels.kernelmanager." + request.param]
```