@mwakaba2 2020-11-04
@kevin-bates Since `FileContentsManager` is a subclass of `ContentsManager`, one option is to create  `AsyncFileContentsManager` for asynchronous file operations and test that. Then, I think we'll be able to test both classes in the test_api file. However, it looks like asyncio doesn't support async file operations by default, so we'll need to use something like [aiofiles: file support for asyncio](https://github.com/Tinche/aiofiles). 

Another option would be to create a no-op or simple custom ContentsManager class that inherits from the async classes, and test that. To mimic network IO, the async methods can include `asyncio.sleep` statements.
I think the first option would be a complete solution and can rely on the existing tests, whereas the second will not exactly test real situations. What would you recommend?