@kevin-bates 2020-11-04
Hi Mariko - I think taking the _forward-moving_ branch by implementing `AsyncFileContentsManager` is the best approach.  Thank you.