@mwakaba2 2020-11-12
Hi @kevin-bates, I added the `AsyncFileContentsManager` class and updated tests to test both async and non-async classes. 
Instead of `aiofiles`, I added a new dependency [`anyio`](https://anyio.readthedocs.io/en/latest/index.html) to run some synchronous callables in a thread with [`run_sync_in_worker_thread`](https://anyio.readthedocs.io/en/latest/threads.html#running-a-function-in-a-worker-thread). This is used to run most `os` operations and `nbformat` methods concurrently. The nbformat team is looking into making the api asynchronous (https://github.com/jupyter/nbformat/issues/194), but it looks like that'll take some time to get that feature out, so using `anyio` for now. 
Please let me know if you have any questions and I'd appreciate your review!