@kevin-bates 2020-11-13
Hi @mwakaba2 - thank you for this great update.  I hope to be able to take this for spin next week - I hope that's okay.  Things look real good via a quick scan - just had the one immediate comment (to show I did look at things :smile: ).

Also, we've made some fairly large changes to the pytest landscape.  This will manifest in your additional tests in the form for requiring `fp_` prefix on the "public" fixtures - which I suspect to be fairly isolated, but a merge with master might be helpful.

Thanks for the links to `nbformat`. The additional dependency on `anyio` seems fine, especially since @bollwyvl is on board!