@kevin-bates 2020-11-19
Hi @mwakaba2 - I'm not sure there's precedent for this and reading up on publishing dev packages to pypi (as I'm not familiar with that particular aspect) I'm finding it not a recommended practice.

What I've done in the past is to use pip with a reference to the GH repo@branch, like so:
```
pip install --upgrade git+https://github.com/mwakaba2/jupyter_server@async-contents-api
```
Would that be sufficient?

Btw, I brought up your PR in the today's server meeting and will try to spend time with this.  I apologize for the delay.