@mwakaba2 2020-11-19
Ah ok no problem. Yeah, we're using an in-house tool to manage 3rd party python dependencies so I'm not able to use pip directly, but I think I found a way to test it.
And thanks! I saw the meeting notes. I'm hoping to complete testing this week.