@kevin-bates 2020-11-20
Hi Mariko - I was just going to post that `LargeFileManager` should be included since it's currently the default `contents_manager_class` value.  Thank you for the update.

I was able to take `AsyncFileContentsManager` for a spin yesterday and found the functionality to be fine.  I decided to run some _crude_ tests using `curl` and uploading/fetching [larger file contents (189 MB)](https://github.com/zemirco/sf-city-lots-json/blob/master/citylots.json).  A "test" simply consisted of timing 3 curl commands (each in the background) to simulate 3 "simultaneous" requests.  Larger content was used solely to trigger longer request times and ensure the previous requests hadn't finished prior to the subsequent requests.  I.e., I wasn't as concerned about the request time as I was about getting multiple requests "serviced" simultaneously.

I found that the current implementation didn't exhibit the blocking behavior that I experienced with kernel startup before we implemented async/await support - which surprised me.  I also found that it seems like the current implementation was slightly better in that you could see a "clear winner" amongst the 3 requests (upwards of 50% faster), whereas, with the async implementation, the times tended to be the same across the 3 requests.  I suspect this is because of "too much cooperation" since each `await` is essentially a _pause point_ for that executor.

Since the primary purpose of this is to address high-latent filesystems, I suspect there will be some "wins" there.  I also wonder if we should take a close look at where `awaits` have been introduced and ensure they make sense.

I'm curious if you see similar behavior.  Thank you for working on this.