@kevin-bates 2020-11-20
Hi Mariko, would you mind splitting the PEP440 update out in a separate PR?  This is something I had on my plate to do today (along with an update to [RELEASE.md](https://github.com/jupyter-server/jupyter_server/blob/master/RELEASE.md) to update the dev version following the release tagging).

I think only the `dev` and `post` suffixes use a dot: https://www.python.org/dev/peps/pep-0440/#public-version-identifiers

If you'd rather I do that, that's fine also.