@kevin-bates 2020-11-20
:smile:  Apparently someone also had this on their plate _yesterday_: https://github.com/jupyter-server/jupyter_server/commit/8763b5b98cb745cc11713af20f72a1919ea2d181

I'll go ahead and submit a PR to add the dot and update RELEASE.md.