@jasongrout 2020-11-20
> I'll go ahead and submit a PR to add the dot and update RELEASE.md.

You may remember that we did this sort of thing in https://github.com/jupyter/jupyter_core/blob/master/jupyter_core/version.py