@kevin-bates 2020-11-20
Thanks @jasongrout - I _just_ saw your comment (after submitting #348).  I hadn't forgotten about your change in core, but personally found that to be overly complex.  That said, if others prefer that approach, I'm okay with adopting it.