@mwakaba2 2020-12-01
@kevin-bates Thanks for putting together the test bundle. 
I ran your put/get scripts 5 times to calculate the avg for each scenario:
  * Before (current implementation on master branch)
```
$ jupyter server --ServerApp.contents_manager_class=jupyter_server.services.contents.largefilemanager.FileContentsManager --debug on main branch)
```
  * After (FileContentsManager)
```
$ jupyter server --ServerApp.contents_manager_class=jupyter_server.services.contents.largefilemanager.FileContentsManager --debug
```
  * After (AsyncFileContentsManager)
```
$ jupyter server --ServerApp.contents_manager_class=jupyter_server.services.contents.largefilemanager.AsyncFileContentsManager --debug
```
Here are the results. 
Fastest request out of three "huge" requests
```
Before
---------
PUT
201 PUT /api/contents/contents-test/huge1.json?token=fd559df06790466297db4a89b210b66ef79a104e935ba17d (::1) 604.49ms
201 PUT /api/contents/contents-test/huge2.json?token=fd559df06790466297db4a89b210b66ef79a104e935ba17d (::1) 517.84ms
201 PUT /api/contents/contents-test/huge2.json?token=fd559df06790466297db4a89b210b66ef79a104e935ba17d (::1) 555.70ms
201 PUT /api/contents/contents-test/huge2.json?token=fd559df06790466297db4a89b210b66ef79a104e935ba17d (::1) 490.21ms
201 PUT /api/contents/contents-test/huge2.json?token=fd559df06790466297db4a89b210b66ef79a104e935ba17d (::1) 707.03ms
---------
avg 575.054ms

GET
200 GET /api/contents/contents-test/huge2.json?token=fd559df06790466297db4a89b210b66ef79a104e935ba17d (::1) 941.87ms
200 GET /api/contents/contents-test/huge2.json?token=fd559df06790466297db4a89b210b66ef79a104e935ba17d (::1) 843.39ms
200 GET /api/contents/contents-test/huge1.json?token=fd559df06790466297db4a89b210b66ef79a104e935ba17d (::1) 873.16ms
201 PUT /api/contents/contents-test/huge1.json?token=fd559df06790466297db4a89b210b66ef79a104e935ba17d (::1) 506.65ms
200 GET /api/contents/contents-test/huge2.json?token=fd559df06790466297db4a89b210b66ef79a104e935ba17d (::1) 873.18ms
---------
avg 807.65ms

After (non-async)
-----------------
PUT
201 PUT /api/contents/contents-test/huge1.json?token=b9c8c67f686f6b746cc4479ddbaee06f8526659d0b10b5a1 (::1) 727.68ms
201 PUT /api/contents/contents-test/huge2.json?token=b9c8c67f686f6b746cc4479ddbaee06f8526659d0b10b5a1 (::1) 520.14ms
201 PUT /api/contents/contents-test/huge1.json?token=b9c8c67f686f6b746cc4479ddbaee06f8526659d0b10b5a1 (::1) 494.54ms
201 PUT /api/contents/contents-test/huge2.json?token=b9c8c67f686f6b746cc4479ddbaee06f8526659d0b10b5a1 (::1) 558.08ms
201 PUT /api/contents/contents-test/huge3.json?token=b9c8c67f686f6b746cc4479ddbaee06f8526659d0b10b5a1 (::1) 599.44ms
---------
avg 579.976ms

GET
200 GET /api/contents/contents-test/huge1.json?token=b9c8c67f686f6b746cc4479ddbaee06f8526659d0b10b5a1 (::1) 836.26ms
200 GET /api/contents/contents-test/huge1.json?token=b9c8c67f686f6b746cc4479ddbaee06f8526659d0b10b5a1 (::1) 819.93ms
200 GET /api/contents/contents-test/huge2.json?token=b9c8c67f686f6b746cc4479ddbaee06f8526659d0b10b5a1 (::1) 834.84ms
200 GET /api/contents/contents-test/huge2.json?token=b9c8c67f686f6b746cc4479ddbaee06f8526659d0b10b5a1 (::1) 857.20ms
200 GET /api/contents/contents-test/huge2.json?token=b9c8c67f686f6b746cc4479ddbaee06f8526659d0b10b5a1 (::1) 828.08ms
---------
avg 835.26ms

After (async)
-----------------
PUT
201 PUT /api/contents/contents-test/huge1.json?token=207ea8b331c25affbb8e607138e46fff0c587935d83cd3f4 (::1) 1253.11ms
201 PUT /api/contents/contents-test/huge2.json?token=207ea8b331c25affbb8e607138e46fff0c587935d83cd3f4 (::1) 893.71ms
201 PUT /api/contents/contents-test/huge3.json?token=207ea8b331c25affbb8e607138e46fff0c587935d83cd3f4 (::1) 917.52ms
201 PUT /api/contents/contents-test/huge1.json?token=207ea8b331c25affbb8e607138e46fff0c587935d83cd3f4 (::1) 1457.99ms
201 PUT /api/contents/contents-test/huge2.json?token=207ea8b331c25affbb8e607138e46fff0c587935d83cd3f4 (::1) 1363.17ms
---------
avg 1177.1ms

GET
200 GET /api/contents/contents-test/huge1.json?token=207ea8b331c25affbb8e607138e46fff0c587935d83cd3f4 (::1) 859.31ms
200 GET /api/contents/contents-test/huge1.json?token=207ea8b331c25affbb8e607138e46fff0c587935d83cd3f4 (::1) 851.68ms
200 GET /api/contents/contents-test/huge2.json?token=207ea8b331c25affbb8e607138e46fff0c587935d83cd3f4 (::1) 876.50ms
200 GET /api/contents/contents-test/huge2.json?token=207ea8b331c25affbb8e607138e46fff0c587935d83cd3f4 (::1) 886.85ms
200 GET /api/contents/contents-test/huge3.json?token=207ea8b331c25affbb8e607138e46fff0c587935d83cd3f4 (::1) 859.66ms
---------
avg 866.8ms
```
I confirmed that there's no noticeable performance difference between "Before" and "After (FileContentsManager)", so that's good. 😀
As for the asynchronous implementation, as you saw, it looks like there's a 104.69% increase in response time for the fastest "huge" PUT request. The increase is minimal for the GET request.

So I removed some await statements to reduce the PUT response time. The increase is minimal now.
```
After (async, optimal)
-----------------
PUT
201 PUT /api/contents/contents-test/huge1.json?token=4f24b1be4ce39fdfd8d911b737f718e1572e4c1bc2dd7c93 (::1) 747.23ms
201 PUT /api/contents/contents-test/huge1.json?token=4f24b1be4ce39fdfd8d911b737f718e1572e4c1bc2dd7c93 (::1) 512.65ms
201 PUT /api/contents/contents-test/huge2.json?token=4f24b1be4ce39fdfd8d911b737f718e1572e4c1bc2dd7c93 (::1) 587.67ms
201 PUT /api/contents/contents-test/huge2.json?token=4f24b1be4ce39fdfd8d911b737f718e1572e4c1bc2dd7c93 (::1) 576.43ms
201 PUT /api/contents/contents-test/huge3.json?token=4f24b1be4ce39fdfd8d911b737f718e1572e4c1bc2dd7c93 (::1) 538.99ms
---------
avg 592.59ms
```
This is just an update for PUT, so I'll look into optimizations for the other file operations next (e.g. rename, delete).